Working Qwixx Alpha

+++ Dice.cpp
@@ -3,7 +3,9 @@
 #include "Dice.h"
 
 std::uniform_int_distribution<int> RandomDice::die{1,6};
-std::default_random_engine RandomDice::generator{static_cast<unsigned long>(time(NULL))};
+
+std::default_random_engine RandomDice::generator{ static_cast<unsigned int>(time(NULL)) };
+//std::default_random_engine RandomDice::generator{static_cast<unsigned long>(time(NULL))};
 
 Dice::Dice(Color _color) : d_color(_color)
 {
